% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiam.r
\name{mind}
\alias{mind}
\title{subsets matrices, keeping the dimensionality intact}
\usage{
mind(x, rows = NULL, cols = NULL)
}
\arguments{
\item{x}{the matrix or data frame to subset}

\item{rows}{either a vector of row indicies or a logical vector giving which rows to select from \code{x}}

\item{cols}{either a vector of column indicies or a logical vector giving which rows to select from \code{x}}
}
\value{
a matrix subsetted to include only the rows and columns indicated but with its dimensionality preserved.
}
\description{
When you subset a matrix or data frame and the condition ends up only grabbing one row, R returns a vector instead of a matrix with one row.  This function returns a matrix with one row.
}
