% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kiam.r
\name{vdim}
\alias{vdim}
\title{return the dimension of a vector x as c(length(x),1)}
\usage{
vdim(x)
}
\arguments{
\item{x}{the object to find the dimension of.  Does exactly what vanilla R does if x is not a vector.  If x is a vector, returns c(length(x),1).}
}
\description{
return the dimension of a vector x as c(length(x),1)
}
